requirements.txt
# requirements.txt

flask==2.2.3
flask-sqlalchemy==3.0.3
apscheduler==3.10.1
alpaca-trade-api==3.0.0
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
ta==0.10.2
python-dotenv==1.0.0
plotly==5.13.1
yfinance==0.2.17
pytz==2022.7.1
